Add optional callback for when the page is loaded
The 'onLoadedCallback' is invoked once the page is loaded and the 'pg-loaded' class has been added
to the body. In most cases, this will be done during the call to finish(), but in some cases it will
be delayed until the loading screen itself has finished the "loading in" animation.

NOTE: this callback is issued before the loading screen itself has "finished" animating out and
removing itself from the DOM.